Separate Sieve of E. from prime.cpp
