+++ src/trinerdi/number-theory/prime_test.cpp
@@ -1,34 +1,6 @@
 #include "gtest/gtest.h"
 #include "prime.cpp"
-
-TEST(SIEVE, BelowTwo) {
-	vector<bool> primes = sieve(2);
-	EXPECT_EQ((vector<bool> {false, false, true}), primes);
-}
-
-TEST(SIEVE, SizeIsInclusive) {
-	vector<bool> primes = sieve(3000);
-	EXPECT_EQ(primes.size(), 3001);
-}
-
-TEST(SIEVE, LastIsComposite) {
-	vector<bool> primes = sieve(3000);
-	EXPECT_EQ(primes[3000], false);
-}
-
-TEST(SIEVE, LastIsPrime) {
-	vector<bool> primes = sieve(2999);
-	EXPECT_EQ(primes[2999], true);
-}
-
-TEST(SIEVE, CutEquals) {
-	vector<bool> primes = sieve(3000);
-	vector<bool> bigger = sieve(30000);
-	bigger.resize(3001);
-	EXPECT_EQ(primes, bigger);
-}
-
-
+#include "sieve.cpp" // Consistency testing
 
 TEST(ISPRIME, Negative) {
 	vector<bool> res = {isprime(-1), isprime(-5237), isprime(-15723953), isprime(-2)};
